Renamed BaseCylinder
