+++ src/js/models/carousel_base_cylinder.js
